requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
watchfiles>=0.22.0

openai>=1.0.0
python-dotenv>=1.0.0